clase13

+++ 2015/2015-I/Clase12/Nave/save.php
@@ -5,10 +5,14 @@ $arrContextOptions=array(
         "verify_peer"=>false,
         "verify_peer_name"=>false,
     ),
-);  
-
-$content = file_get_contents($_POST['url'],false, stream_context_create($arrContextOptions));
+);
+$content = file_get_contents($_POST['url'],false, 
+	stream_context_create($arrContextOptions));
 $name = sprintf("assets/%s.jpg",$_POST['id']);
-file_put_contents($name, $content);
+if(!file_exists($name)){
+	file_put_contents($name, $content);
+}
+
+
 
  ?>
